Move jbang test to e2e test module
