--- api-tests/jbang/tx-eval-success.json
@@ -1,17 +1,0 @@
-{
-  "type": "jsonwsp/response",
-  "version": "1.0",
-  "servicename": "ogmios",
-  "methodname": "EvaluateTx",
-  "result": {
-    "EvaluationResult": {
-      "spend:0": {
-        "memory": 500,
-        "steps": 64100
-      }
-    }
-  },
-  "reflection": {
-    "id": "2211864e-8e14-4331-97c8-cfb9c3a49400"
-  }
-}
